refactored map
